requirements-dev.txt
asgiref==3.2.7
astroid==2.4.1
beautifulsoup4==4.9.1
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.6
django-allauth==0.42.0
django-bootstrap4==1.1.1
django-google-api==1.0.1
Flask==1.1.2
google-api-core==1.20.0
google-api-python-client==1.9.3
google-auth==1.16.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
httplib2==0.18.1
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpickle==1.4.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauth2client==4.1.3
oauthlib==3.1.0
protobuf==3.12.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.2
pylint-django==2.0.11
pylint-plugin-utils==0.6
python-decouple==3.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.15.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.3.1
static3==0.7.0
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
